Test pass, added BATCHSIZE to have a bigger input to test on
